requirements.txt
# Core dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
requests>=2.31.0

# File parsing
pdfplumber>=0.10.0
openpyxl>=3.1.0
pandas>=2.0.0

# Output generation
python-docx>=1.0.0
reportlab>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# SQL parsing utilities
sqlparse>=0.4.4
